refactor: dto단 command 패키지 분리
