+++ src/main/java/com/example/temp/fanpool/dto/command/ParticipateFanpoolCommand.java
@@ -1,4 +1,4 @@
-package com.example.temp.fanpool.dto;
+package com.example.temp.fanpool.dto.command;
 
 public record ParticipateFanpoolCommand(
         long userId,
