+++ src/main/java/com/example/temp/fanpool/service/DeleteFanpoolParticipationService.java
@@ -1,6 +1,6 @@
 package com.example.temp.fanpool.service;
 
-import com.example.temp.fanpool.dto.DeleteFanpoolParticipationCommand;
+import com.example.temp.fanpool.dto.command.DeleteFanpoolParticipationCommand;
 
 public interface DeleteFanpoolParticipationService {
     void doService(DeleteFanpoolParticipationCommand command);
