+++ src/main/java/com/example/temp/fanpool/service/FindFilteredFanpoolService.java
@@ -1,8 +1,8 @@
 package com.example.temp.fanpool.service;
 
 import com.example.temp.fanpool.dto.FindFilteredFanpoolResponse;
-import com.example.temp.fanpool.dto.FindFilteredFanpoolCommand;
+import com.example.temp.fanpool.dto.command.FindFilteredFanpoolCommand;
 
 public interface FindFilteredFanpoolService {
-    FindFilteredFanpoolResponse doService(long userId, FindFilteredFanpoolCommand command);
+    FindFilteredFanpoolResponse doService(FindFilteredFanpoolCommand command);
 }
