+++ src/main/java/com/example/temp/fanpool/service/FindHostedFanpoolByUserService.java
@@ -1,6 +1,6 @@
 package com.example.temp.fanpool.service;
 
-import com.example.temp.fanpool.dto.FindHostedFanpoolByUserCommand;
+import com.example.temp.fanpool.dto.command.FindHostedFanpoolByUserCommand;
 import com.example.temp.fanpool.dto.FindHostedFanpoolByUserResponse;
 
 public interface FindHostedFanpoolByUserService {
