+++ src/main/java/com/example/temp/fanpool/service/ParticipateFanpoolService.java
@@ -1,6 +1,6 @@
 package com.example.temp.fanpool.service;
 
-import com.example.temp.fanpool.dto.ParticipateFanpoolCommand;
+import com.example.temp.fanpool.dto.command.ParticipateFanpoolCommand;
 
 public interface ParticipateFanpoolService {
     void doService(ParticipateFanpoolCommand command);
