+++ src/main/java/com/example/temp/fanpool/service/UpdateFanpoolService.java
@@ -1,6 +1,6 @@
 package com.example.temp.fanpool.service;
 
-import com.example.temp.fanpool.dto.UpdateFanpoolCommand;
+import com.example.temp.fanpool.dto.command.UpdateFanpoolCommand;
 
 public interface UpdateFanpoolService {
     void doService(UpdateFanpoolCommand command);
