+++ src/main/java/com/example/temp/fanpool/service/UpdateFanpoolStateService.java
@@ -1,6 +1,6 @@
 package com.example.temp.fanpool.service;
 
-import com.example.temp.fanpool.dto.UpdateFanpoolStateCommand;
+import com.example.temp.fanpool.dto.command.UpdateFanpoolStateCommand;
 
 public interface UpdateFanpoolStateService {
     void doService(UpdateFanpoolStateCommand command);
